Do not needlessly use "long" types.
git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@35331 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ sub/vobsub.c
@@ -837,7 +837,7 @@ int vobsub_parse_ifo(void* this, const char *const name, unsigned int *palette,
         } else if (memcmp(block, ifo_magic, strlen(ifo_magic) + 1))
             mp_msg(MSGT_VOBSUB, MSGL_ERR, "VobSub: Bad magic in IFO header\n");
         else {
-            unsigned long pgci_sector = block[0xcc] << 24 | block[0xcd] << 16
+            unsigned pgci_sector = block[0xcc] << 24 | block[0xcd] << 16
                 | block[0xce] << 8 | block[0xcf];
             int standard = (block[0x200] & 0x30) >> 4;
             int resolution = (block[0x201] & 0x0c) >> 2;
@@ -870,8 +870,8 @@ int vobsub_parse_ifo(void* this, const char *const name, unsigned int *palette,
                 || rar_read(block, sizeof(block), 1, fd) != 1)
                 mp_msg(MSGT_VOBSUB, MSGL_ERR, "VobSub: Can't read IFO PGCI\n");
             else {
-                unsigned long idx;
-                unsigned long pgc_offset = block[0xc] << 24 | block[0xd] << 16
+                unsigned idx;
+                unsigned pgc_offset = block[0xc] << 24 | block[0xd] << 16
                     | block[0xe] << 8 | block[0xf];
                 for (idx = 0; idx < 16; ++idx) {
                     unsigned char *p = block + pgc_offset + 0xa4 + 4 * idx;
